Tightens up the mini registry submission process
