+++ package.json
@@ -38,7 +38,7 @@
     "@nuxtjs/robots": "^5.2.10",
     "@octokit/core": "^7.0.2",
     "@tailwindcss/vite": "^4.1.8",
-    "@types/node": "^22.15.30",
+    "@types/node": "^24.0.0",
     "@types/youtube": "^0.1.2",
     "@vercel/analytics": "^1.5.0",
     "@vercel/speed-insights": "^1.2.0",
@@ -53,7 +53,7 @@
     "outdent": "^0.8.0",
     "pg": "^8.16.0",
     "redirect-ssl": "^3.0.0",
-    "sass": "^1.89.1",
+    "sass": "^1.89.2",
     "tailwindcss": "^4.1.8",
     "uuid": "^11.1.0",
     "vue-cookie-law": "^1.13.3"
